pyproject.toml
[project]
name = "content_agent"
version = "0.1.0"
description = "content_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "python-dotenv",
    "openai",
    "langchain-community",
    "beautifulsoup4",
    "selenium"
]

[project.scripts]
content_agent = "content_agent.main:run"
run_crew = "content_agent.main:run"
train = "content_agent.main:train"
replay = "content_agent.main:replay"
test = "content_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
